Refactor: DB 모델 모듈 생성에 따른 모듈 이름 변경
+ 미사용 모듈 임포트 삭제

Related to: #4
